requirements.txt
streamlit>=1.41.0
pandas>=2.2.0
numpy>=1.26.0
openai>=1.60.0
anthropic>=0.40.0
google-generativeai>=0.8.0
fpdf2>=2.8.0
